Moved login helpers to spec/support/
